removing unwanted code
