--- packages/geoview-core/src/core/components/legend/legend-context.ts
@@ -1,4 +1,0 @@
-import { createContext } from 'react';
-
-// eslint-disable-next-line @typescript-eslint/no-empty-function
-export const LegendContext = createContext({ selectedLayer: null, setSelectedLayer: () => {} });
